+++ CHANGES.rst
@@ -2,6 +2,8 @@ v 1.6
 =====
 
 * Added a ``display_time`` keyword to the ``flash()`` method.
+* Changed the ``animate`` property to a method so that it can take
+  a ``stop_time`` keyword to control the amount of time between loop steps.
 
 v 1.5
 =====